requirements.txt
colorama==0.3.7
decorator==4.0.11
dnspython==1.15.0
numpy==1.12.1
pathlib2==2.1.0
psutil==5.0.1
pycrypto==2.6.1
six==1.10.0
gunicorn==19.7.1
Flask==1.0.2
requests==2.16.5
python-telegram-bot==12.0.0
Pillow==6.1.0